Fix: image erosion now works at borders
previously when we tried to erode images to prevent black line artifacts at the
perimeter, the erosion doesn't work due to the MATLAB convention of erosion of
positive values at the border of the image. Now we fix this by first padding
the image with zeros, erode, then remove the zeros that we added.